Backend/effbrainy/requirements.txt
Django==1.8.5
Pygments==2.0.2
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==3.2.4
gunicorn==19.3.0
psycopg2==2.6.1
static3==0.6.1
wsgiref==0.1.2
django-twilio==0.4
httplib2==0.8
six==1.4.1
twilio==3.6.3
unittest2==0.5.1
django-oauth-toolkit==0.7.2
django-oauth2-provider==0.2.4
oauthlib==0.7.2
requests==2.8.0
cloudinary==1.1.3
Pillow == 2.9.0